backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
sqlmodel==0.0.22
psycopg[binary]
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings>=2.7.0
pydantic>=2.12.0